+++ index.js
@@ -1,5 +1,4 @@
 /* eslint-disable import/no-unresolved */
 require('./lib/apiActionSteps.js');
 require('./lib/apiVerificationSteps.js');
-require('./lib/hooks.js');
 require('./lib/parameterTypes.js');